+++ package.json
@@ -13,35 +13,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^0.4.5",
-    "@angular/animations": "^5.2.8",
-    "@angular/common": "^5.2.8",
-    "@angular/compiler": "^5.2.8",
-    "@angular/core": "^5.2.8",
-    "@angular/forms": "^5.2.8",
-    "@angular/http": "^5.2.8",
-    "@angular/platform-browser": "^5.2.8",
-    "@angular/platform-browser-dynamic": "^5.2.8",
-    "@angular/router": "^5.2.8",
-    "core-js": "^2.4.1",
-    "file-saver": "^1.3.3",
-    "omggif": "^1.0.9",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.14"
+    "@angular-devkit/core": "^0.4.9",
+    "@angular/animations": "^5.2.9",
+    "@angular/cdk": "^5.2.4",
+    "@angular/common": "^5.2.9",
+    "@angular/compiler": "^5.2.9",
+    "@angular/core": "^5.2.9",
+    "@angular/forms": "^5.2.9",
+    "@angular/http": "^5.2.9",
+    "@angular/material": "^5.2.4",
+    "@angular/platform-browser": "^5.2.9",
+    "@angular/platform-browser-dynamic": "^5.2.9",
+    "@angular/router": "^5.2.9",
+    "core-js": "^2.5.4",
+    "file-saver": "^1.3.8",
+    "rxjs": "^5.5.8",
+    "zone.js": "^0.8.25"
   },
   "devDependencies": {
     "@angular/cli": "1.6.2",
-    "@angular/compiler-cli": "^5.2.8",
-    "@angular/language-service": "^5.2.8",
+    "@angular/compiler-cli": "^5.2.9",
+    "@angular/language-service": "^5.2.9",
     "@types/file-saver": "^1.3.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.102",
+    "@types/node": "^6.0.104",
     "autoprefixer": "^6.5.3",
     "circular-dependency-plugin": "^4.2.1",
     "codelyzer": "^4.2.1",
-    "copy-webpack-plugin": "^4.5.0",
-    "css-loader": "^0.28.10",
+    "copy-webpack-plugin": "^4.5.1",
+    "css-loader": "^0.28.11",
     "cssnano": "^3.10.0",
     "exports-loader": "^0.6.3",
     "file-loader": "^1.1.11",
@@ -56,8 +57,8 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "less-loader": "^4.1.0",
-    "postcss-loader": "^2.1.1",
-    "postcss-url": "^7.3.1",
+    "postcss-loader": "^2.1.3",
+    "postcss-url": "^7.3.2",
     "raw-loader": "^0.5.1",
     "sass-loader": "^6.0.7",
     "source-map-loader": "^0.2.0",
