+++ src/app/app.component.html
@@ -1,7 +1,5 @@
 <!--The content below is only a placeholder and can be replaced.-->
-<div style="text-align:center">
-  <h1>
-    Welcome to {{ title }}!
-  </h1>
-</div>
+<mat-toolbar>
+  <span>Save your lols</span>
+</mat-toolbar>
 <app-render-lols></app-render-lols>
