+++ src/app/app.component.scss
@@ -1,0 +1,1 @@
+@import '~@angular/material/prebuilt-themes/deeppurple-amber.css';
