+++ src/app/app.module.ts
@@ -1,11 +1,11 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { MatToolbarModule, MatGridListModule } from '@angular/material';
 
 import { AppComponent } from './app.component';
 import { RenderLolsComponent } from './render-lols/render-lols.component';
 import { FetchLolService } from './fetch-lol.service';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { FetchLolServiceMock } from './fetch-lol.service.mock';
 
 
@@ -16,7 +16,9 @@ import { FetchLolServiceMock } from './fetch-lol.service.mock';
   ],
   imports: [
     BrowserModule,
-    HttpClientModule
+    HttpClientModule,
+    MatToolbarModule,
+    MatGridListModule
   ],
   providers: [
     {provide: FetchLolService, useClass: FetchLolServiceMock}
