+++ src/app/fetch-lol.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { LolData } from './lolData.model';
+import { LolData } from './render-lols/lolData.model';
 import { Observable } from 'rxjs/Observable';
 // import { map } from 'rxjs/operators';
 import 'rxjs/add/operator/map';
@@ -10,10 +10,6 @@ import 'rxjs/add/operator/do';
 export class FetchLolService {
 
   constructor(private httpClient: HttpClient) { }
-
-  renderGif() {
-    return this.httpClient.get('assets/example.gif');
-  }
 
   fetchLol(): Observable<LolData> {
     console.log('Start fetching');
