+++ src/app/render-lols/render-lols.component.ts
@@ -1,14 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { LolData } from '../lolData.model';
+import { LolData } from './lolData.model';
 import { FetchLolService } from '../fetch-lol.service';
 import { Observable } from 'rxjs/Observable';
-import { gen_anim } from './gif-service';
 import * as FileSaver from 'file-saver';
 
 @Component({
   selector: 'app-render-lols',
   templateUrl: './render-lols.component.html',
-  styleUrls: ['./render-lols.component.css']
+  styleUrls: ['./render-lols.component.scss']
 })
 export class RenderLolsComponent implements OnInit {
 
@@ -22,11 +21,6 @@ export class RenderLolsComponent implements OnInit {
 
   fetchGif() {
     this.lolData$ = this.fetchLolService.fetchLol();
-  }
-
-  saveGif(buffer) {
-    console.log(buffer);
-    FileSaver.saveAs(new Blob(gen_anim(buffer)), './_tmp.gif');
   }
 
 }
